requirements.txt
Flask
python-dotenv
requests
gunicorn
psycopg2-binary
